requirements.txt
docx==0.2.4
pandas==1.4.2
PyPDF2==2.11.1
python-dotenv==0.21.0
python_docx==0.8.11
selenium==4.5.0
streamlit==1.13.0
xlsxwriter==3.0.3